challenge_3/requirements.txt
azure-functions
azure-cosmos==4.0.0b6